requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
pillow>=9.5.0
opencv-python>=4.7.0

# Machine Learning & Metrics
scikit-learn>=1.2.0
scipy>=1.10.0

# Visualization & Logging
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.12.0
tqdm>=4.65.0

# Configuration
pyyaml>=6.0

# Medical Imaging (Optional)
# pydicom>=2.3.0
# nibabel>=5.1.0

# Utilities
h5py>=3.8.0
